+++ warden/api/routes/dependencies/authorized_users.py
@@ -5,11 +5,11 @@ from fastapi import Depends, FastAPI, Request
 from warden.lib.config import APIConfig
 
 
-def init_authorized_users(app: FastAPI, api_config: APIConfig):
-    app.state.authorized_users = api_config.authorized_users
+def init_authorized_users(app: FastAPI, authorized_users: list[str]):
+    app.state.authorized_users = authorized_users
 
 
-def get_authorized_users(request: Request) -> APIConfig:
+def get_authorized_users(request: Request) -> list[str]:
     conf = getattr(request.app.state, "authorized_users", None)
     if conf is None:
         raise RuntimeError(
