+++ warden/api/routes/sessions.py
@@ -7,7 +7,7 @@ from sqlalchemy import Select
 
 from warden.api.routes.dependencies.auth import verify_root
 from warden.api.routes.dependencies.db import DBSessionDep
-from warden.api.routes.dependencies.users import AuthorizedUsersDep
+from warden.api.routes.dependencies.authorized_users import AuthorizedUsersDep
 from warden.api.schemas.sessions import CreateSession, SessionResponse
 from warden.lib.models.sessions import Session
